Move headers and redirects parsers into workers-shared
