+++ packages/workers-shared/utils/configuration/parseRedirects.ts
